requirements.txt
flask
python-kafka
dataset
paho-mqtt
psycopg2
chartkick
